refactor create infraestructure monitor page
